+++ packages/ingest/granule.js
@@ -1,18 +1,18 @@
 'use strict';
 
+const crypto = require('crypto');
 const deprecate = require('depd')('my-module');
 const fs = require('fs-extra');
 const cloneDeep = require('lodash.clonedeep');
-const get = require('lodash.get');
 const groupBy = require('lodash.groupby');
 const identity = require('lodash.identity');
+const get = require('lodash.get');
 const omit = require('lodash.omit');
 const os = require('os');
 const path = require('path');
 const urljoin = require('url-join');
 const encodeurl = require('encodeurl');
 const cksum = require('cksum');
-const checksum = require('checksum');
 const xml2js = require('xml2js');
 const { aws, log } = require('@cumulus/common');
 const errors = require('@cumulus/common/errors');
@@ -40,21 +40,19 @@ class Discover {
       throw new TypeError('Can not construct abstract class.');
     }
 
-    const config = get(event, 'config');
-
-    this.buckets = get(config, 'buckets');
-    this.collection = get(config, 'collection');
-    this.provider = get(config, 'provider');
-    this.useList = get(config, 'useList');
+    this.buckets = event.config.buckets;
+    this.collection = event.config.collection;
+    this.provider = event.config.provider;
+    this.useList = event.config.useList;
     this.event = event;
 
-    this.port = get(this.provider, 'port', 21);
-    this.host = get(this.provider, 'host', null);
-    this.path = get(this.collection, 'provider_path') || '/';
+    this.port = this.provider.port || 21;
+    this.host = this.provider.host;
+    this.path = this.collection.provider_path || '/';
 
     this.endpoint = urljoin(this.host, this.path);
-    this.username = get(this.provider, 'username', null);
-    this.password = get(this.provider, 'password', null);
+    this.username = this.provider.username;
+    this.password = this.provider.password;
 
     // create hash with file regex as key
     this.regexes = {};
@@ -123,7 +121,7 @@ class Discover {
       aws.s3ObjectExists({ Bucket: discoveredFile.bucket, Key: discoveredFile.name })
         .then((exists) => (exists ? null : discoveredFile)))))
       .filter(identity);
-    
+
     // Group the files by granuleId
     const filesByGranuleId = groupBy(newFiles, (file) => file.granuleId);
 
@@ -174,7 +172,7 @@ class Granule {
     this.collection = collection;
     this.provider = provider;
 
-    if (this.collection ) this.collection.url_path = this.collection.url_path || '';
+    if (this.collection) this.collection.url_path = this.collection.url_path || '';
     this.port = get(this.provider, 'port', 21);
     this.host = get(this.provider, 'host', null);
     this.username = get(this.provider, 'username', null);
@@ -377,13 +375,13 @@ class Granule {
   * @returns {Promise} checksum value calculated from file
   **/
   async _hash(algorithm, filepath) {
-    const options = { algorithm };
-
-    return new Promise((resolve, reject) =>
-      checksum.file(filepath, options, (err, sum) => {
-        if (err) return reject(err);
-        return resolve(sum);
-      }));
+    return new Promise((resolve, reject) => {
+      const hash = crypto.createHash(algorithm);
+      const fileStream = fs.createReadStream(filepath);
+      fileStream.on('error', reject);
+      fileStream.on('data', (chunk) => hash.update(chunk));
+      fileStream.on('end', () => resolve(hash.digest('hex')));
+    });
   }
 
   /**
@@ -691,7 +689,7 @@ async function updateMetadata(granuleId, cmrFile, files, distEndpoint, published
   // add/replace the OnlineAccessUrls
   const metadata = await getMetadata(cmrFile.filename);
   const metadataObject = await parseXmlString(metadata);
-  const metadataGranule = get(metadataObject, 'Granule');
+  const metadataGranule = metadataObject.Granule;
   const updatedGranule = {};
   Object.keys(metadataGranule).forEach((key) => {
     if (key === 'OnlineResources' || key === 'Orderable') {